+++ packages/timeline-state-resolver/src/conductor.ts
@@ -21,6 +21,7 @@ import {
 	DeviceOptionsBase,
 	TimelineDatastoreReferences,
 	Datastore,
+	DeviceOptionsTelemetrics,
 } from 'timeline-state-resolver-types'
 
 import { DoOnTime } from './devices/doOnTime'
@@ -48,6 +49,7 @@ import { VMixDevice, DeviceOptionsVMixInternal } from './integrations/vmix'
 import { OBSDevice, DeviceOptionsOBSInternal } from './integrations/obs'
 import { VizMSEDevice, DeviceOptionsVizMSEInternal } from './integrations/vizMSE'
 import { ShotokuDevice, DeviceOptionsShotokuInternal } from './integrations/shotoku'
+import { TelemetricsDevice } from './devices/telemetrics'
 
 export { DeviceContainer }
 export { CommandWithContext }
@@ -579,6 +581,15 @@ export class Conductor extends EventEmitter<ConductorEvents> {
 					getCurrentTime,
 					threadedClassOptions
 				)
+			} else if (deviceOptions.type === DeviceType.TELEMETRICS) {
+				newDevice = await DeviceContainer.create<DeviceOptionsTelemetrics, typeof TelemetricsDevice>(
+					'../../dist/devices/telemetrics.js',
+					'TelemetricsDevice',
+					deviceId,
+					deviceOptions,
+					getCurrentTime,
+					threadedClassOptions
+				)
 			} else {
 				// @ts-ignore deviceOptions.type is of type "never"
 				const type: any = deviceOptions.type
@@ -663,7 +674,6 @@ export class Conductor extends EventEmitter<ConductorEvents> {
 				this.emit('warning', 'Error when terminating device', e)
 			}
 			await device.terminate()
-
 			this.devices.delete(deviceId)
 		} else {
 			return Promise.reject('No device found')
@@ -1456,6 +1466,7 @@ export type DeviceOptionsAnyInternal =
 	| DeviceOptionsVMixInternal
 	| DeviceOptionsShotokuInternal
 	| DeviceOptionsVizMSEInternal
+	| DeviceOptionsTelemetrics
 
 function removeParentFromState(o: TimelineState): TimelineState {
 	for (const key in o) {